Creation of the SchoolHelperService bean and its implementation
